Code/qgai/face/requirements.txt
pillow~=11.3.0
opencv-python~=4.12.0.88
numpy~=2.0.2
joblib~=1.5.1
scikit-learn~=1.6.1
insightface~=0.7.3